refactor: swapped dashboard and create incident pages around
